[RAPTOR-3008] split artifact predictors into separate files (#46)
